// frontend/dvf-map/src/app/models/dvf-property.model.ts
export interface DvfProperty {
  id_mutation?: string
  date_mutation?: string
  valeur_fonciere?: number
  type_local?: string
  latitude?: number
  longitude?: number
  adresse_numero?: string
  adresse_nom_voie?: string
  code_postal?: string
  nom_commune?: string
  surface?: number
  surface_terrain?: number
  id_parcelle?: string
}